# openapi.yaml
openapi: 3.0.3
info:
  title: FPL Assistant API
  description: >
    Backend API for the FPL Assistant website and MCP server (see
    PRD-fpl-assistant.md and TDD-fpl-assistant.md). Covers team lookup, player
    stats/search/comparison, fixtures (with double/blank gameweek detection),
    squad view, captain/transfer suggestions, mini-league standings and rival
    comparison, and chart data.

    Draft v0.1 — pending review (see tasks/todo.md Task 4). Not yet locked.
  version: 1.5.0
  contact:
    name: FPL Assistant
    email: jeffvu304@gmail.com
servers:
  - url: https://api.fplassistant.example.com
    description: Production (placeholder host, per TDD's deferred hosting decision)
...
paths:
  /teams/{teamId}:
    get:
      operationId: getTeam
      tags: [Team]
      summary: Look up a public FPL team by team ID
      description: >
        No auth — team ID is validated against the cached bootstrap-static
        entry list. Returns a clean 404 if the team ID doesn't exist.
      parameters:
        - $ref: '#/components/parameters/TeamId'
      responses:
        '200':
          description: Team found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
              example:
                team_id: 1234567
                name: "Fantasy Wanderers"
                manager_name: "A. Manager"
                overall_rank: 154032
                total_points: 412
                data_as_of: "2026-09-06T09:00:00Z"
                stale: false
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
  /teams/{teamId}/suggest-transfer:
    get:
      operationId: suggestTransfer
      tags: [Recommendations]
      summary: Suggest a transfer for the given team
      description: >
        Opinionated tool — returns the shared RecommendationContract shape
        (TDD-fpl-assistant.md worked example: `recommendation` is
        `{ player_out, player_in }`). `risk_tolerance` is a per-query
        parameter, not stored state (no accounts in MVP).
      parameters:
        - $ref: '#/components/parameters/TeamId'
        - name: risk_tolerance
          in: query
          description: Defaults to "balanced"
          schema:
            type: string
            enum: [safe, balanced, differential]
            default: balanced
      responses:
        '200':
          description: Transfer recommendation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecommendationContract'
              example:
                recommendation:
                  player_out:
                    player_id: 210
                    name: "J. Struggling"
                  player_in:
                    player_id: 415
                    name: "B. Differential"
                confidence: "medium"
                reasoning:
                  - factor: "underlying_stats"
                    detail: "player_in leads the league in xG per 90 among sub-10%-owned midfielders"
                    sample_size: 8
                  - factor: "fixture_run"
                    detail: "player_in's team has 3 fixtures rated 2 or easier in the next 5 gameweeks"
                    sample_size: 5
                alternatives:
                  - option:
                      player_out:
                        player_id: 210
                        name: "J. Struggling"
                      player_in:
                        player_id: 322
                        name: "C. Template"
                    reasoning:
                      - factor: "ownership"
                        detail: "Owned by 45% of top 10k managers"
                        sample_size: 10
                    why_not_top_pick: "Safer but offers no differential rank upside"
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
  /teams/{teamId}/compare/{rivalTeamId}:
    get:
      operationId: compareRivalTeam
      tags: [MiniLeague]
      summary: Compare a team against a rival team (e.g. a mini-league opponent)
      description: Returns both teams' summary data side by side for head-to-head comparison.
      parameters:
        - $ref: '#/components/parameters/TeamId'
        - name: rivalTeamId
          in: path
          required: true
          description: Public FPL team ID of the rival to compare against
          schema:
            type: integer
      responses:
        '200':
          description: Comparison result
          content:
            application/json:
              schema:
                type: object
                properties:
                  team:
                    $ref: '#/components/schemas/Team'
                  rival_team:
                    $ref: '#/components/schemas/Team'
                  data_as_of:
                    $ref: '#/components/schemas/DataAsOf'
                  stale:
                    $ref: '#/components/schemas/Stale'
              example:
                team:
                  team_id: 1234567
                  name: "Fantasy Wanderers"
                  manager_name: "A. Manager"
                  overall_rank: 154032
                  total_points: 412
                rival_team:
                  team_id: 7654321
                  name: "Rival FC"
                  manager_name: "B. Rival"
                  overall_rank: 201044
                  total_points: 398
                data_as_of: "2026-09-06T09:00:00Z"
                stale: false
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  parameters:
    TeamId:
      name: teamId
      in: path
      required: true
      description: Public FPL team ID
      schema:
        type: integer
...
  responses:
    NotFound:
      description: The requested ID doesn't exist (e.g. invalid/unknown team ID)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: "That team ID doesn't exist"
    TooManyRequests:
      description: Rate limit exceeded (60 requests/minute/IP)
      headers:
        Retry-After:
          description: Seconds to wait before retrying
          schema:
            type: string
            pattern: '^[0-9]+$'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: "Rate limit exceeded, please retry shortly"
    ServiceUnavailable:
      description: Recommendation couldn't be produced because required data is unavailable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/DegradedState'
          example:
            message: "Unable to produce a recommendati
            on right now, please retry shortly"
  schemas:
    DataAsOf:
      type: string
      format: date-time
      description: Timestamp of the underlying cached data this response was served from
    Stale:
      type: boolean
      description: >
        True if this response was served from a cached value because the live
        FPL API was unreachable (graceful degradation, not an error state)
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
    DegradedState:
      type: object
      required: [message]
      properties:
        message:
          type: string
    Team:
      type: object
      properties:
        team_id:
          type: integer
        name:
          type: string
        manager_name:
          type: string
        overall_rank:
          type: integer
        total_points:
          type: integer
        data_as_of:
          $ref: '#/components/schemas/DataAsOf'
        stale:
          $ref: '#/components/schemas/Stale'
...
    RecommendationContract:
      type: object
      required: [recommendation, confidence, reasoning, alternatives]
      properties:
        recommendation:
          type: object
          description: Shape varies per tool (see suggest_captain / suggest_transfer examples)
          additionalProperties: true
        confidence:
          type: string
          enum: [low, medium, high]
          description: >
            Derived from sample size of the relevant signal: fewer than 3
            relevant gameweeks -> low, 3-10 -> medium, 10+ -> high. Reflects
            signal strength only, never a claim of correctness.
        reasoning:
          type: array
          items:
            $ref: '#/components/schemas/ReasoningFactor'
        alternatives:
          type: array
          items:
            $ref: '#/components/schemas/Alternative'
        data_as_of:
          type: string
          format: date-time
          nullable: true
          description: >
            Present only when this recommendation was served from stale
            cached data; the timestamp the underlying data was last
            successfully fetched.
    ReasoningFactor:
      type: object
      required: [factor, detail, sample_size]
      properties:
        factor:
          type: string
        detail:
          type: string
        sample_size:
          type: integer
    Alternative:
      type: object
      required: [option, reasoning, why_not_top_pick]
      properties:
        option:
          type: object
          description: Shape matches the sibling `recommendation` field for the same tool
          additionalProperties: true
        reasoning:
          type: array
          items:
            $ref: '#/components/schemas/ReasoningFactor'
        why_not_top_pick:
          type: string